+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Show icon for preview link in list tables next to the state.
 * Change interface in block editor to match latest editor design.
 * Update sidebar description to include the preview link.
+* Extend Preview dropdown for public preview in WordPress 6.7+.
 
 ## 2.10.0 (2022-11-19)
 * Compatibility with WordPress 6.1.